services/agent-generator/requirements-base.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
portkey-ai>=1.10.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3
structlog>=24.1.0
